+++ Sources/KeyboardKit/Appearance/View+Button.swift
@@ -14,6 +14,7 @@ public extension View {
      Apply the standard button style, including a background,
      foreground color, corner radius, shadow and font.
      */
+    @available(*, deprecated, message: "Take an appearance provider instead of context")
     func standardButtonStyle(
         for action: KeyboardAction,
         context: KeyboardContext,
@@ -28,6 +29,7 @@ public extension View {
     /**
      Apply a standard button background.
      */
+    @available(*, deprecated, message: "Move this to the appearance provider instead")
     func standardButtonBackground(
         for action: KeyboardAction,
         context: KeyboardContext) -> some View {
@@ -37,10 +39,11 @@ public extension View {
     /**
      Apply a standard button font.
      */
+    @available(*, deprecated, message: "Move this to the appearance provider instead")
     func standardButtonFont(
         for action: KeyboardAction,
         context: KeyboardContext) -> some View {
-        let provider = context.keyboardAppearanceProvider
+        let provider = keyboardInputViewController.keyboardAppearanceProvider
         let rawFont = Font(provider.font(for: action))
         if let weight = provider.fontWeight(for: action, context: context) { return font(rawFont.weight(weight)) }
         let hasImage = action.standardButtonImage(for: context) != nil
@@ -50,6 +53,7 @@ public extension View {
     /**
      Apply a standard button foreground color.
      */
+    @available(*, deprecated, message: "Move this to the appearance provider instead")
     func standardButtonForeground(for context: KeyboardContext) -> some View {
         foregroundColor(.standardButtonTint(for: context))
     }
@@ -57,6 +61,7 @@ public extension View {
     /**
      Apply a standard button shadow.
      */
+    @available(*, deprecated, message: "Move this to the appearance provider instead")
     func standardButtonShadow(for context: KeyboardContext) -> some View {
         let color = Color.standardButtonShadowColor(for: context)
         return self.shadow(color: color, radius: 0, x: 0, y: 1)