+++ Sources/KeyboardKit/Fakes/FakeKeyboardContext.swift
@@ -19,7 +19,6 @@ public class FakeKeyboardContext: KeyboardContext {
     public var emojiCategory: EmojiCategory = .smileys
 
     public var actionHandler: KeyboardActionHandler = FakeKeyboardActionHandler()
-    public var keyboardAppearanceProvider: KeyboardAppearanceProvider = StandardKeyboardAppearanceProvider()
     public var keyboardInputSetProvider: KeyboardInputSetProvider = StandardKeyboardInputSetProvider()
     public var keyboardLayoutProvider: KeyboardLayoutProvider = StandardKeyboardLayoutProvider()